overlay: ensure that existing phandles are not overwritten
A phandle in an overlay is not supposed to overwrite a phandle that
already exists in the base dtb as this breaks references to the
respective node in the base.

Traditionally all phandle values in the overlay were increased by a
fixed offset such that the base and overlay handle values don't overlap.
With this change the overlay's phandles are determined dynamically.
Already existing phandles are kept and an array "phandlemap" is used to
track how the overlay's phandles change.

overlay_adjust_local_phandles() which creates the mapping between old
and new phandles in the overlay needs to check the nodes in the base. To
be able to find the base's node, overlay_fixup_phandles() must be run
first, so the order of functions fdt_overlay_apply() had to change.

A test is added that checks that newly added phandles and existing
phandles work as expected.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>