move release version into VERSION.txt
To synchronize the release version of the Makefile and Meson build
systems, pull the version info from a shared file.

Meson requires that the shared library version follow the X.Y.Z
numbering scheme. But the Makefile supported building shared libraries
with an EXTRAVERSION appended, for example X.Y.Z-rc1. We want to keep
the output of both systems the same to avoid issues, so we drop support
for the Makefile EXTRAVERSION.

Signed-off-by: Brandon Maier <brandon.maier@collins.com>