version_get.h: add missing info in Meson build
The Makefile calls the setlocalversion script to get the repo info,
append it to the version, and add it to version_gen.h. Meson is skipping
the repo info and just writing the version into version_gen.h.

The Meson vcs_tag() can call an external program to get the version, but
it can't append the output to another string, so we must update
setlocalversion to do it for us.

Signed-off-by: Brandon Maier <brandon.maier@collins.com>