libfdt: allow building only static or shared libraries
When using the meson-python backend to compile libfdt, it throws an
installation error that it does not know where to install static
libraries. We would disable the static libraries, but Meson is always
building and installing both the shared and static libraries.

Right now we always build both shared and static libraries because our
'static-build' option needs us to link the dtc binaries against the
static libfdt. So we need to make sure static is always built even if
the build is configured for shared only.

So instead lets not link against the static library but link all of
libfdt's objects directly. We will always have access to the library
objects regardless if the build is configured for static or shared.

Signed-off-by: Brandon Maier <brandon.maier@collins.com>