setup: Add dependencies for Windows
Windows apparently builds pylibfdt if swig is declared as a dependency
in the setup file. Otherwise it must be installed manually and Windows
does not sport a package manager, so this is hard for users.

This change seems to have no ill effects on Linux, at least.

So add a new setup_requires line.

Commit-notes:
This was picked up from:

https://github.com/devicetree-org/pylibfdt/pull/2
END

Series-to: dtc
Series-cc: Gua Guo <gua.guo@intel.com>, david

Signed-off-by: Gua Guo <gua.guo@intel.com>
Signed-off-by: Simon Glass <sjg@chromium.org>