setup: Add dependencies for Windows
Windows apparently builds pylibfdt if swig is declared as a dependency
in the setup file. Otherwise it must be installed manually and Windows
does not sport a package manager, so this is hard for users.

This change seems to have no ill effects on Linux, at least.

So add a new setup_requires line.

Signed-off-by: Gua Guo <gua.guo@intel.com>
Signed-off-by: Simon Glass <sjg@chromium.org>
Message-ID: <20240811150248.7537-6-sjg@chromium.org>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>