Restore phandle references from __fixups__ node
The __fixups__ node contains information about labels. Parse its
properties to create phandle markers which improve the resulting dts
when decompiling a device tree blob.

Note that this is essential to do before the __fixups__ node is
removed in generate_fixups_tree() because otherwise the information
contained in that node is lost.

Fixes: 915daadbb62d ("Start with empty __local_fixups__ and __fixups__ nodes")
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@baylibre.com>