fdtput: Fix documentation about existing nodes
The documentation claims that `-c` would "Create nodes if they don't
already exist". This is true, but suggests that trying to create a node
that already exists is not an error. fdtput however errors out in that
case. Similar `fdtput -d` errors out when called for a non-existing
node.

Drop the "if they [don't] already exist" to make that clearer.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@baylibre.com>