tests: Add compatibility with uutils
In some places run_tsets.sh needs to get the size of files, which it does
with stat(1).  However the syntax to do this is different between GNU
coreutils stat(1) and BSD's stat(1).  We have some logic that looks for
"GNU" in the version string to figure out the correct version.

This will break upcoming Ubuntu versions which are now using uutils, a Rust
reimplementation of coreutils.  These support the same GNU syntax, but
don't have the "GNU" in the version string.

Update the detection to simply try the GNU version and otherwise assume
BSD.

Link: https://github.com/dgibson/dtc/issues/166

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>