fdtput: Use strtol() in preference to sscanf()
I find the exact semantics of scanf() always confusing, and mostly we use
strtol() and similar functions instead.  Switch fdtput to using strtol and
similar functions instead of sscanf().

As well as being more in keeping with what we do in other places, this
works around a FreeBSD bug[0], where sscanf(), but not strtol() parses
"003" as having value 0.

[0] https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=288440

Link: https://github.com/dgibson/dtc/issues/165

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>