libfdt: Improve size savings in FDT_RO_PROBE slightly
In the case where we have set FDT_ASSUME_MASK to disable
ASSUME_VALID_DTB checks, we can improve the FDT_RO_PROBE macro slightly.
The first thing that fdt_ro_probe_() does when we can_assume(VALID_DTB)
is true is to return whatever the contents of the totalsize field of the
DTB is. Since the FDT_RO_PROBE macro only cares about a negative value
there, we can optimize this check such that we are to assume it's a
valid DTB, we don't need to do anything here.

Signed-off-by: Tom Rini <trini@konsulko.com>
Message-ID: <20251210022002.3004223-3-trini@konsulko.com>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>