libfdt: fdt_get_name: Add can_assume(VALID_DTB) check
In this function from fdt_ro.c we have (reasonably) some checks of the
DTB before we begin work. However, we do this in a way that we cannot
make use of the normal FDT_RO_PROBE macro and instead have a direct call
to fdt_ro_probe_(). Add a test for !can_assume(VALID_DTB) here first so
that in cases where we are assuming a valid DTB we can omit the checks.

Signed-off-by: Tom Rini <trini@konsulko.com>
Message-ID: <20251210022002.3004223-4-trini@konsulko.com>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>